+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   [💎] Renaming `Custom Marker Icon URL` to `Custom marker URL` to follow Obsidian guidelines
 -   [💎] Showing a toast notice if the user uses an invalid combination of maki icon and custom marker url
+-   [💎] Showing a notice if the plugin has updated so that the user may check for new features
 
 ## 1.1.0